Updated .md file frontmatter
Now all the .md files reference a tags JSON file (e.g., project-1-tags references content/tags/project-1-tags.json). Also updated the title and author of each project to reflect one in the spreadsheet, in column order. Left the numerical numbering for now, to make it easy to track which tags go with which project