Update FilterMenu.jsx
Added an icon for the "close" button at the top of the filter menu. Changed the positioning from absolute to fixed (relative to browser window) at small screens, static at medium and larger. Added a button to apply filters at the bottom of the menu (closes the menu). Added styling to the reset and apply filter buttons.