Refactored to use <Image/> component
ProjectCard now takes <Image> as a child. All the images are now stored in the projects content collection and can be referenced dynamically. I will note it hasn't helped with the lag in loading the images on the project cards, so it might not be worth it. Maybe the images can be stored in the public folder and referenced from there. I'm not sure what's more intuitive to someone making a pull request to update their project.