Update githubApiHelper.js
Refactored so that the async function getContributors passes the data coming back from the async GitHub call to the getMostRecentContributors function within itself. This requires that the data come back before getMostRecentContributors can be called, which is where the error keeps occurring that the data can't be found in time for the forEach function.