Top-level folder now contains OSSI site; changed GitHub token access to JaneliaSciComp
Deleted the sibling folder containing the aborted attempt using the Astrowind template, so now the OSSI site is the only and top-level folder. This also makes the OSSI site README the primary readme for the repo.

In this commit, I also changed the GitHub token access to a personal access token I made to access the metadata of this repo in JaneliaSciComp. This token is stored in a .env file that I have not uploaded to the remote repository.